+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.3.2 (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v1.3.1...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [1.3.1 (2024-09-18)](https://github.com/kdeldycke/extra-platforms/compare/v1.3.0...v1.3.1)
 
 - Fix conflicting detection heuristics for Linux distributions. Closes #72.